Added informative output to stderr, because this information is visible in the OUTPUT window of vsCode
These messages may be printed to log.debug, if there is a way to make the debug output visible when running
fortls as a plugin in vsCode (via the Modern Fortran plugin). Or perhaps it is not necessary at all.
It helped me in finding out what was wrong in fortls/my settings.